Move `getPageOptions` to `config.js`
