+++ lib/config.js
@@ -109,3 +109,18 @@ module.exports.getPuppeteerLaunchConfig = () => {
     executablePath: puppeteerChromiumExecutable || null
   };
 };
+
+module.exports.getPageOptions = printSize => {
+  if (printSize) {
+    const dimensions = printSize.match(/^([\d.]+)x([\d.]+)([a-z]*)$/);
+    if (dimensions) {
+      const [width, height, unit] = dimensions.slice(1);
+      return { width: `${width}${unit}`, height: `${height}${unit}` };
+    }
+    return { format: printSize };
+  } else if (revealConfig.width && revealConfig.height) {
+    return { width: revealConfig.width, height: revealConfig.height };
+  } else {
+    return { format: 'A4' };
+  }
+};
