+++ lib/print.js
@@ -2,7 +2,7 @@
 const path = require('path');
 const debug = require('debug')('reveal-md');
 const { revealBasePath } = require('./constants');
-const { getRevealOptions, getPuppeteerLaunchConfig } = require('./config');
+const { getPuppeteerLaunchConfig, getPageOptions } = require('./config');
 let puppeteer;
 
 try {
@@ -12,28 +12,12 @@ try {
   debug(err);
 }
 
-const getPageOptions = (revealOptions, printSize) => {
-  if (printSize) {
-    const dimensions = printSize.match(/^([\d.]+)x([\d.]+)([a-z]*)$/);
-    if (dimensions) {
-      const [width, height, unit] = dimensions.slice(1);
-      return { width: `${width}${unit}`, height: `${height}${unit}` };
-    }
-    return { format: printSize };
-  } else if (revealOptions.width && revealOptions.height) {
-    return { width: revealOptions.width, height: revealOptions.height };
-  } else {
-    return { format: 'A4' };
-  }
-};
-
 module.exports = async (initialUrl, print, printSize) => {
   if (!puppeteer) {
     return;
   }
 
   const puppeteerLaunchConfig = getPuppeteerLaunchConfig();
-  const revealOptions = getRevealOptions();
 
   const printPluginPath = path.join(revealBasePath, 'plugin', 'print-pdf', 'print-pdf.js');
 
@@ -49,9 +33,7 @@ module.exports = async (initialUrl, print, printSize) => {
     const page = await browser.newPage();
 
     const pdfOptions = { path: pdfFilename, printBackground: true };
-    Object.assign(pdfOptions, getPageOptions(revealOptions, printSize));
-
-    console.log(pdfOptions);
+    Object.assign(pdfOptions, getPageOptions(printSize));
 
     await page.goto(`${initialUrl}?print-pdf`, { waitUntil: 'load' });
     await page.pdf(pdfOptions);
